requirements.txt
easyocr
Flask
Flask-Cors
Pillow
ultralytics
gunicorn
psycopg2-binary